Complete translation to Spanish
Hi, Marek. Here is a complete translation to Spanish. We are using your application in Vitalinux EDU DGA, a version of Lubuntu in order to use in the public educational system of Aragon (a region of the Northern Spain). My name, for the credits of translatiors, is José M. Ciordia. Thanks and best regards.